+++ packages/article-comments/src/styles/index.js
@@ -1,8 +1,14 @@
 import { StyleSheet } from "react-native";
+import { fonts, fontSizes } from "@times-components/styleguide";
 import sharedStyles from "./shared";
 
 const styles = StyleSheet.create({
-  ...sharedStyles
+  ...sharedStyles,
+  link: {
+    ...sharedStyles.link,
+    fontFamily: fonts.supporting,
+    fontSize: fontSizes.commentsGuidelines
+  }
 });
 
 export default styles;