+++ packages/article-magazine-comment/__tests__/web/semantic.web.test.js
@@ -147,7 +147,7 @@ const tests = [
             onTopicPress={() => {}}
             onVideoPress={() => {}}
             receiveChildList={() => {}}
-            spotAccountId=""
+            commentingConfig={{ account: { current: "dummiy-spotim-id" } }}
           />
         </ContextProviderWithDefaults>
       );