+++ packages/article-skeleton/__tests__/web/article-with-isPreview.web.test.js
@@ -310,7 +310,7 @@ const renderArticle = () => (
       onTopicPress={() => {}}
       onTwitterLinkPress={() => {}}
       onVideoPress={() => {}}
-      spotAccountId=""
+      commentingConfig={{ account: { current: "dummiy-spotim-id" } }}
       isPreview={isPreview}
       inlineRelatedArticlesFlag
     />