+++ packages/article-skeleton/src/article-body/article-body.web.js
@@ -56,7 +56,8 @@ import {
   NativeAd,
   NativeAdTitle,
   Ad,
-  InlineAdWrapper
+  InlineAdWrapper,
+  InlineAdTitle
 } from "../styles/article-body/responsive";
 
 const deckApiUrl = "https://gobble.timesdev.tools/deck/api/deck-post-action/";
@@ -100,6 +101,7 @@ const renderers = ({
   ad(key) {
     return (
       <InlineAdWrapper className="hidden">
+        <InlineAdTitle>Advertisement</InlineAdTitle>
         <AdContainer key={key} slotName="inline-ad" />
       </InlineAdWrapper>
     );
@@ -107,6 +109,7 @@ const renderers = ({
   inlineAd1(key) {
     return (
       <InlineAdWrapper className="hidden">
+        <InlineAdTitle>Advertisement</InlineAdTitle>
         <AdContainer key={key} slotName="inlineAd1" />
       </InlineAdWrapper>
     );
@@ -114,6 +117,7 @@ const renderers = ({
   inlineAd2(key) {
     return (
       <InlineAdWrapper className="hidden">
+        <InlineAdTitle>Advertisement</InlineAdTitle>
         <AdContainer key={key} slotName="inlineAd2" />
       </InlineAdWrapper>
     );
@@ -121,6 +125,7 @@ const renderers = ({
   inlineAd3(key) {
     return (
       <InlineAdWrapper className="hidden">
+        <InlineAdTitle>Advertisement</InlineAdTitle>
         <AdContainer key={key} slotName="inlineAd3" />
       </InlineAdWrapper>
     );