+++ packages/article-paragraph/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.222](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.4.221...@times-components/article-paragraph@1.4.222) (2021-09-02)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
+## [1.4.221](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.4.220...@times-components/article-paragraph@1.4.221) (2021-09-01)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
+## [1.4.220](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.4.219...@times-components/article-paragraph@1.4.220) (2021-09-01)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
+## [1.4.219](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.4.218...@times-components/article-paragraph@1.4.219) (2021-09-01)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
+## [1.4.218](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.4.217...@times-components/article-paragraph@1.4.218) (2021-08-31)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
 ## [1.4.217](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.4.216...@times-components/article-paragraph@1.4.217) (2021-08-27)
 
 **Note:** Version bump only for package @times-components/article-paragraph