+++ packages/article/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.8.4](https://github.com/newsuk/times-components/compare/@times-components/article@7.8.3...@times-components/article@7.8.4) (2021-09-02)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.8.3](https://github.com/newsuk/times-components/compare/@times-components/article@7.8.2...@times-components/article@7.8.3) (2021-09-01)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.8.2](https://github.com/newsuk/times-components/compare/@times-components/article@7.8.1...@times-components/article@7.8.2) (2021-09-01)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.8.1](https://github.com/newsuk/times-components/compare/@times-components/article@7.8.0...@times-components/article@7.8.1) (2021-09-01)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+# [7.8.0](https://github.com/newsuk/times-components/compare/@times-components/article@7.7.65...@times-components/article@7.8.0) (2021-08-31)
+
+
+### Features
+
+* **TE-178:**  real name commenting ([#2780](https://github.com/newsuk/times-components/issues/2780)) ([4a51ffe](https://github.com/newsuk/times-components/commit/4a51ffefb0b37720f9328f0baaa940e1083c87ec))
+
+
+
+
+
 ## [7.7.65](https://github.com/newsuk/times-components/compare/@times-components/article@7.7.64...@times-components/article@7.7.65) (2021-08-27)
 
 **Note:** Version bump only for package @times-components/article