+++ packages/message-bar/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.67](https://github.com/newsuk/times-components/compare/@times-components/message-bar@0.3.66...@times-components/message-bar@0.3.67) (2021-08-31)
+
+**Note:** Version bump only for package @times-components/message-bar
+
+
+
+
+
 ## [0.3.66](https://github.com/newsuk/times-components/compare/@times-components/message-bar@0.3.65...@times-components/message-bar@0.3.66) (2021-08-09)
 
 **Note:** Version bump only for package @times-components/message-bar